rename old WES-specific orchestrator
