+++ tests/test_wes_orchestrator.py
@@ -7,11 +7,11 @@ from bravado.client import SwaggerClient, ResourceDecorator
 from bravado.testing.response_mocks import BravadoResponseMock
 
 from ga4ghtest.services.wes.controller import WESService
-from ga4ghtest.core.orchestrator import run_job
-from ga4ghtest.core.orchestrator import run_submission
-from ga4ghtest.core.orchestrator import run_queue
-from ga4ghtest.core.orchestrator import monitor_queue
-from ga4ghtest.core.orchestrator import monitor
+from ga4ghtest.core.wes_orchestrator import run_job
+from ga4ghtest.core.wes_orchestrator import run_submission
+from ga4ghtest.core.wes_orchestrator import run_queue
+from ga4ghtest.core.wes_orchestrator import monitor_queue
+from ga4ghtest.core.wes_orchestrator import monitor
 
 
 def test_run_job(mock_queue_config,
@@ -19,17 +19,17 @@ def test_run_job(mock_queue_config,
                  mock_submission, 
                  mock_wes, 
                  monkeypatch):
-    monkeypatch.setattr('ga4ghtest.core.orchestrator.queue_config',
+    monkeypatch.setattr('ga4ghtest.core.wes_orchestrator.queue_config',
                         lambda: mock_queue_config)
-    monkeypatch.setattr('ga4ghtest.core.orchestrator.wes_config',
+    monkeypatch.setattr('ga4ghtest.core.wes_orchestrator.wes_config',
                         lambda: mock_wes_config)
-    monkeypatch.setattr('ga4ghtest.core.orchestrator.fetch_queue_workflow', 
+    monkeypatch.setattr('ga4ghtest.core.wes_orchestrator.fetch_queue_workflow', 
                         lambda x: mock_queue_config[x])
-    monkeypatch.setattr('ga4ghtest.core.orchestrator.create_submission', 
+    monkeypatch.setattr('ga4ghtest.core.wes_orchestrator.create_submission', 
                         lambda **kwargs: None)
-    monkeypatch.setattr('ga4ghtest.core.orchestrator.WESService',
+    monkeypatch.setattr('ga4ghtest.core.wes_orchestrator.WESService',
                         lambda wes_id: mock_wes)
-    monkeypatch.setattr('ga4ghtest.core.orchestrator.update_submission', 
+    monkeypatch.setattr('ga4ghtest.core.wes_orchestrator.update_submission', 
                         lambda w,x,y,z: None)
 
     mock_request = {'workflow_url': None,
@@ -51,12 +51,12 @@ def test_run_submission(mock_submission,
                         mock_run_log,
                         mock_wes, 
                         monkeypatch):
-    monkeypatch.setattr('ga4ghtest.core.orchestrator.get_submission_bundle', 
+    monkeypatch.setattr('ga4ghtest.core.wes_orchestrator.get_submission_bundle', 
                         lambda x,y: mock_submission['mock_sub'])
-    monkeypatch.setattr('ga4ghtest.core.orchestrator.update_submission', 
+    monkeypatch.setattr('ga4ghtest.core.wes_orchestrator.update_submission', 
                         lambda w,x,y,z: None)
 
-    monkeypatch.setattr('ga4ghtest.core.orchestrator.run_job', 
+    monkeypatch.setattr('ga4ghtest.core.wes_orchestrator.run_job', 
                         lambda **kwargs: mock_run_log)
 
     test_run_log = run_submission(queue_id='mock_queue',
@@ -69,15 +69,15 @@ def test_run_queue(mock_queue_config,
                    mock_submission,
                    mock_queue_log,
                    monkeypatch):
-    monkeypatch.setattr('ga4ghtest.core.orchestrator.queue_config', 
+    monkeypatch.setattr('ga4ghtest.core.wes_orchestrator.queue_config', 
                         lambda: mock_queue_config)
-    monkeypatch.setattr('ga4ghtest.core.orchestrator.get_submissions', 
+    monkeypatch.setattr('ga4ghtest.core.wes_orchestrator.get_submissions', 
                         lambda x,status: ['mock_sub'])
-    monkeypatch.setattr('ga4ghtest.core.orchestrator.get_submission_bundle', 
+    monkeypatch.setattr('ga4ghtest.core.wes_orchestrator.get_submission_bundle', 
                         lambda x,y: mock_submission['mock_sub'])
 
     mock_run_log = mock_submission['mock_sub']['run_log']
-    monkeypatch.setattr('ga4ghtest.core.orchestrator.run_submission', 
+    monkeypatch.setattr('ga4ghtest.core.wes_orchestrator.run_submission', 
                         lambda **kwargs: mock_run_log)
 
     test_queue_log = run_queue(queue_id='mock_queue_1', 
@@ -93,17 +93,17 @@ def test_monitor_queue(mock_submission,
                        mock_queue_log, 
                        mock_wes, 
                        monkeypatch):
-    monkeypatch.setattr('ga4ghtest.core.orchestrator.get_submissions', 
+    monkeypatch.setattr('ga4ghtest.core.wes_orchestrator.get_submissions', 
                         lambda **kwargs: ['mock_sub'])
-    monkeypatch.setattr('ga4ghtest.core.orchestrator.get_submission_bundle', 
+    monkeypatch.setattr('ga4ghtest.core.wes_orchestrator.get_submission_bundle', 
                         lambda x,y: mock_submission['mock_sub'])
-    monkeypatch.setattr('ga4ghtest.core.orchestrator.WESService', 
+    monkeypatch.setattr('ga4ghtest.core.wes_orchestrator.WESService', 
                         lambda wes_id: mock_wes)
-    monkeypatch.setattr('ga4ghtest.core.orchestrator.convert_timedelta', 
+    monkeypatch.setattr('ga4ghtest.core.wes_orchestrator.convert_timedelta', 
                         lambda x: 0)
-    monkeypatch.setattr('ga4ghtest.core.orchestrator.ctime2datetime', 
+    monkeypatch.setattr('ga4ghtest.core.wes_orchestrator.ctime2datetime', 
                         lambda x: dt.datetime.now())
-    monkeypatch.setattr('ga4ghtest.core.orchestrator.update_submission', 
+    monkeypatch.setattr('ga4ghtest.core.wes_orchestrator.update_submission', 
                         lambda w,x,y,z: None)
 
     mock_wes.get_run_status.return_value = {'run_id': 'mock_run',
